prevent subtract with overflow

+++ server/service/src/rnr_form/insert.rs
@@ -180,7 +180,7 @@ fn validate(
         }
 
         // If the the previous form was from many periods ago, don't use it in subsequent calculations
-        if this_period - previous_period > 1 {
+        if this_period as isize - previous_period as isize > 1 {
             return Ok((None, period, master_list_id));
         }
     }
